==> 01_preprocessing_pseudobulk_h1299.ipynb <==
# ! pip install scrublet

import scanpy as sc
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import warnings
import scanpy.external as sce
import sys
import scipy 

https://satijalab.org/seurat/archive/v2.4/conversion_vignette.html 
перевод адата в сеурат объект


https://doi.org/10.1016/j.isci.2021.102151
Wyler et al. (2021) iScience

Ссылка:  https://cellxgene.cziscience.com/collections/d0e9c47b-4ce7-4f84-b182-eddcfa0b2658


#### extraction of viral genes: https://github.com/BIMSBbioinfo/Ewyler_SARS-CoV/blob/master/Processing/Process_Seurat.rmd


#### methods: https://www.cell.com/cms/10.1016/j.isci.2021.102151/attachment/12909dee-01ca-466f-86f9-3e997d6461dd/mmc1

#!curl -o h1299.h5ad "https://corpora-data-prod.s3.amazonaws.com/db9bf174-33a3-451e-9997-51f834b8975b/local.h5ad?AWSAccessKeyId=ASIATLYQ5N5XZWH3EPDD&Signature=iUhg%2BDsx%2FmhhmdGKq6WakRhaCRc%3D&x-amz-security-token=IQoJb3JpZ2luX2VjEGQaCXVzLXdlc3QtMiJGMEQCIAn71ESlNGsut3yrcxCwL%2Fm1NqZumhYj1DETreLQ0CwsAiA59IHI1aK8SBKOPyXk19DoRGyAhu%2Bfhv5Y4zGM8Hk94SrrAwgdEAEaDDIzMTQyNjg0NjU3NSIM6Gpytff0VHlC%2BUGMKsgD%2BUq%2FLB3dL67QzbaMYLrAZz%2Fr28yxC9KLvDTt0KVCto8TTEG60%2Fkh9RekJfhV680M7nblfa6jjXVNMnCWUSBb%2BzVeCVvShw5N7WQAdw2HzQALmlj%2FtG52dwgztVulW1epxYA5wYwCVx6qVCGpEBtdlLzOVVAGeAXGXh%2Fq0rPvMDbCeuV5%2B7hhHn%2Ft3GMhZNAMbSOb3kXMuyhdUULrDwgMDO6p7O8WuN2WhOcqKiq5k7%2F1wgbOjyN4L8HvWgQSypyqSAJ228jztvqB8%2BSkRCXzOkTL%2BsxX4dgBf2QXb7GHvJmUXhhW9klo1Dq1dNhNGNw59d9yoDt60qDFs6YLYxe2DzguDszXJP2iUb7n3NCU%2FWUwNFtRTsxZYPnYieWdL8GT1ICzkJ9IFKA0LohILe4vB%2F2Eo9h4373Y%2BYrJxaFb5D65czKxgZIq5EvTVUsm9tStWyH10hT%2Fulz2eX2Y8%2BKqcfj4T99l8JlStPXAZUCPfop2FMICnt%2Fm07H4aPLur2EQy%2FMgch0cK3BX26tFKC31m%2B%2FaBHIxr53O5RvzBd7gmhj7N2zrspFQG%2Ff%2BBeVP%2FsrgIh0AxnpC0qduiQBFeUG8Gav1tVcg8P4yMIWImaAGOqYBk2kLLXOH%2BiEyzRPplbWKXts0jG50OuzJ5o2Xh9BVjgTqLOM3lqJdefE0GdvXi8VEKCztBWQY3SJJOWxq0sGuPB%2FONypNxKXmdXvJzCDn2ywQcpD38DBK2xp2fCZ8EORt02VJXqMOt3IoWBaqPJkKIwuSSflkHmHEo8Si%2FK5i6GoXbY56zxdqz%2Fbdd3tpdmw3i4G04IyFoqjW%2FTsLuNOTaX%2FuEEw1bQ%3D%3D&Expires=1678746466"

adata = sc.read_h5ad('h1299.h5ad')

### remove S1 infected samples  and check what experimental conditions are available

adata = adata[adata.obs["strain"] != "SARSCoV1"]
adata = adata[adata.obs.query("strain != 'SARSCoV1'").index]
adata = adata[[r != "SARSCoV1" for r in adata.obs.strain]]
adata.obs.strain.unique()

conditions = []
for i in range(len(adata.obs.index)):
    conditions.append(adata.obs.index[i].split('-')[1:3])
set(tuple(row) for row in conditions)

### return raw count values

# adata.layers["counts"] = adata.X.copy()

adata.X = adata.raw.X

### Filter low quality cells

adata.var["mito"] = adata.var.feature_name.str.startswith("MT-")
adata.var[adata.var.mito==True] 

fig, axes = plt.subplots(ncols = 3, nrows = 1, figsize=(15,4))


fig.suptitle('QC')

sns.histplot(adata.obs['nCount_RNA'], bins=50, ax=axes[1])
plt.xlim(-1000, 60000)


sns.histplot(adata.obs['n_genes_by_counts'], bins=25, ax=axes[0])
plt.xlim(0, 6500)

sns.histplot(adata.var['mito'], bins=10, ax=axes[2])
plt.xlim(-0.3, 1.5)

### remove doublets 

# filter cells and genes
sc.pp.filter_cells(adata, min_genes=200)
sc.pp.filter_genes(adata, min_cells=3) 

sce.pp.scrublet(adata, mean_center=True)
sce.pp.scrublet_simulate_doublets(adata)
adata

adata = adata[~adata.obs.predicted_doublet]

genes = list(adata.var['feature_name'].values)

### matrix with genes and their expression levels in corresponding cells

counts = adata[:, adata.var['feature_name'].isin(genes)].to_df() 

counts.to_csv('preprocessed_h1299.csv')

#  starting with prepocessed counts

import scanpy as sc
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import warnings
import scanpy.external as sce
import sys
import scipy 

counts = pd.read_csv('preprocessed_h1299.csv', sep=',')

counts.head()

counts4 = counts[counts['Unnamed: 0'].str.contains('-4')]
counts4.index = counts4['Unnamed: 0']
counts4.drop('Unnamed: 0', axis = 1)
counts4 = counts4.groupby([s.split('_')[0] for s in counts4.index.values]).sum().T
counts4.to_csv('h1299_4h_sum.csv')
counts4

counts12 = counts[counts['Unnamed: 0'].str.contains('-12')]
counts12.index = counts12['Unnamed: 0']
counts12.drop('Unnamed: 0', axis = 1)
counts12 = counts12.groupby([s.split('_')[0] for s in counts12.index.values]).sum().T
counts12.to_csv('h1299_12h_sum.csv')
counts12

counts24 = counts[counts['Unnamed: 0'].str.contains('24')]
counts24.index = counts24['Unnamed: 0']
counts24.drop('Unnamed: 0', axis = 1)
counts24 = counts24.groupby([s.split('_')[0] for s in counts24.index.values]).sum().T
counts24.to_csv('h1299_24h_sum.csv')
counts24

counts36 = counts[counts['Unnamed: 0'].str.contains('36')]
counts36.index = counts36['Unnamed: 0']
counts36.drop('Unnamed: 0', axis = 1)
counts36 = counts36.groupby([s.split('_')[0] for s in counts36.index.values]).sum().T
counts36.to_csv('h1299_36h_sum.csv')
counts36

# DIFFERENTIAL EXPRESSION CASE VS CONTROL


import rpy2.robjects as ro
from rpy2.robjects.packages import importr
from rpy2.robjects import pandas2ri
from rpy2.robjects.conversion import localconverter

pandas2ri.activate()
DESeq2 = importr("DESeq2")

from rpy2.robjects import default_converter
from rpy2.robjects.conversion import rpy2py
base = importr("base")

# for non-summed matrices

def prep_df(filename):
    df = pd.read_csv(filename, sep=',')
    genes = list(df.columns[1:])
    counts = df.T
    counts.columns = df.iloc[:, 0]
    counts = counts.drop('Unnamed: 0', axis=0)
    counts = counts.reset_index()
    counts = counts.drop('index', axis=1)

    return counts, genes

def deseq(meta: pd.DataFrame, counts: pd.DataFrame, formula: str, ref: str, exp: str):
    # Calculate normalization factors
    dds = DESeq2.DESeqDataSetFromMatrix(
        countData=counts, colData=meta, design=ro.Formula(formula))
    
    dds = DESeq2.DESeq(dds) #parallel=True
    #estimateSizeFactors(dds, type = 'iterate')
    
    print(f"experiment_{exp}_vs_{ref}")
    resR = DESeq2.results(dds, name=f"experiment_{ref}_vs_{exp}")
    res = r_to_df(resR)
    res = res.sort_values("padj")
    res = res.loc[res["padj"] < 0.05]
    res = res.loc[res["log2FoldChange"].abs() > 0.01]

    return res

def r_to_df(r_df):
    with localconverter(default_converter + pandas2ri.converter):
        return rpy2py(base.as_data_frame(r_df))

counts = counts36

design = pd.DataFrame({
    "experiment": [sample_name.split("-")[1] for sample_name in counts.columns],
    #"hpi": [sample_name.split("-")[1] for sample_name in counts.columns],
}, index=counts.columns)

design

res = deseq(meta=design, counts=counts, formula="~experiment", ref="mock", exp="S2")
res
