==> 12_DEsingle_prep_h1299.ipynb <==
#  starting with prepocessed counts
# create raw count matrices with 4h as control and other hrs as tests and check the dimensionality of each group to create group factor in R

import scanpy as sc
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import warnings
import scanpy.external as sce
import sys
import scipy 

counts = pd.read_csv('preprocessed_h1299.csv', sep=',')

genes = list(counts.columns[1:])

conditions = []
for i in range(len(counts['Unnamed: 0'])):
    conditions.append(counts['Unnamed: 0'][i].split('-')[1:3])
set(tuple(row) for row in conditions)

counts4 = counts[counts['Unnamed: 0'].str.contains('-4')]
counts4.index = counts4['Unnamed: 0']
counts4 = counts4.rename(columns={'Unnamed: 0': "genes"})
counts4.index = counts4.genes
counts4 = counts4.drop('genes', axis = 1)
counts4 = counts4.T

counts12 = counts[counts['Unnamed: 0'].str.contains('12h')]
counts12.index = counts12['Unnamed: 0']
counts12 = counts12.rename(columns={'Unnamed: 0': "genes"})
counts12.index = counts12.genes
counts12 = counts12.drop('genes', axis = 1)
counts12 = counts12.T

counts_12_contr = pd.merge(counts12, counts4, on=counts12.index)
counts_12_contr.index=counts_12_contr['key_0'].values
counts_12_contr = counts_12_contr.drop('key_0', axis = 1)
counts_12_contr.to_csv('h1299_12hpi.csv')

print('exp:', counts_12_contr.filter(regex='-S2').shape[1], 'mock:', counts_12_contr.filter(regex='mock').shape[1])

counts4

counts24 = counts[counts['Unnamed: 0'].str.contains('24')]
counts24.index = counts24['Unnamed: 0']
counts24 = counts24.rename(columns={'Unnamed: 0': "genes"})
counts24.index = counts24.genes
counts24 = counts24.drop('genes', axis = 1)
counts24 = counts24.T

counts_24_contr = pd.merge(counts24, counts4, on=counts24.index)
counts_24_contr.index=counts_24_contr['key_0'].values
counts_24_contr = counts_24_contr.drop('key_0', axis = 1)
counts_24_contr.to_csv('h1299_24hpi.csv')

counts_24_contr

counts_24_contr.filter(regex='-S2').shape[1]

counts_24_contr.filter(regex='mock').shape[1]

#exp <- 8780 # the number of S2 samples in the columns, H1299 24 HPI
#mock <- 4261 # the number of mock samples in the columns, H1299 24 HPI

counts36 = counts[counts['Unnamed: 0'].str.contains('S2-36')]
counts36.index = counts36['Unnamed: 0']
counts36 = counts36.rename(columns={'Unnamed: 0': "genes"})
counts36.index = counts36.genes
counts36 = counts36.drop('genes', axis = 1)
counts36 = counts36.T

counts_36_contr = pd.merge(counts36, counts4, on=counts36.index)
counts_36_contr.index=counts_36_contr['key_0'].values
counts_36_contr = counts_36_contr.drop('key_0', axis = 1)
counts_36_contr.to_csv('h1299_36hpi.csv')

counts_36_contr.filter(regex='-S2').shape[1]

counts_36_contr.filter(regex='mock').shape[1]

#exp <- 9382 # the number of S2 samples in the columns, H1299 36 HPI
#mock <- 4261 # the number of mock samples in the columns, H1299 36 HPI

counts_36_contr
