==> 14_desingle_h1299.ipynb <==
if(!require(BiocManager)) 
  install.packages("BiocManager")

# mkdir -p ~/local/R_libs

BiocManager::install("DEsingle", lib="~/local/R_libs/")
BiocManager::install("BiocParallel", lib="~/local/R_libs/")

library(DEsingle, lib="~/local/R_libs/")
library(BiocParallel, lib="~/local/R_libs/")

counts <- read.csv(file = 'h1299_36hpi.csv')

rownames(counts) <- counts$X
head(counts)

drop <- c("X")
counts = counts[,!(names(counts) %in% drop)]

head(counts)

# exp <- 11558 # the number of S2 samples in the columns at 8 hpi in CALU3
# mock <- 9829 # the number of mock samples in the columns at 8 hpi in CALU3

#exp <- 4685 # the number of S2 samples in the columns at 12 hpi in CALU3
#mock <- 6683 # the number of mock samples in the columns at 12 hpi in CALU3

# exp <- 12125 # the number of S2 samples in the columns at 12 hpi in H1299
# mock <- 4261 # the number of mock samples in the columns at 12 hpi in H1299

# exp <- 8780 # the number of S2 samples in the columns at 24 hpi in H1299
# mock <- 4261 # the number of mock samples in the columns at 24 hpi in H1299

exp <- 9382 # the number of S2 samples in the columns at 36 hpi in H1299
mock <- 4261 # the number of mock samples in the columns at 36 hpi in H1299




group <- factor(c(rep(1,exp), rep(2,mock)))

# Set the parameters and register the back-end to be used
param <- MulticoreParam(workers = 16, progressbar = TRUE)
register(param)

counts <- na.omit(counts)

# Detecting the DE genes
results <- DEsingle(counts = counts, group = group, parallel = TRUE, BPPARAM = param) # , parallel = TRUE, BPPARAM = param

# Dividing the DE genes into 3 categories at threshold of FDR < 0.05
results.classified <- DEtype(results = results, threshold = 0.05)

head(counts)

write.table(results.classified, file = 'desingle_h1299_36h.csv', sep=',')
