==> 23_interscection_of_genes_by_two_methods.ipynb <==
import pandas as pd

! ls

d12 = pd.read_csv('h1299_DEGs_12h.csv')
d24 = pd.read_csv('h1299_DEGs_24h.csv')
d36 = pd.read_csv('h1299_control_4_36.csv')

s12 = pd.read_csv('/Users/sofialuk/Desktop/thesis/h1299/h1299_desingle/desingle_h1299_12h.csv')
s24 = pd.read_csv('/Users/sofialuk/Desktop/thesis/h1299/h1299_desingle/desingle_h1299_24h.csv')
s36 = pd.read_csv('/Users/sofialuk/Desktop/thesis/h1299/h1299_desingle/desingle_h1299_36h.csv')

d12 = d12[(d12['padj'] < 0.05) & (d12['log2FoldChange'] > 0.5)]
d24 = d24[(d24['padj'] < 0.05) & (d24['log2FoldChange'] > 0.5)]
d36 = d36[(d36['padj'] < 0.05) & (d36['log2FoldChange'] > 0.5)]

s12.iloc[:, 10:]

s12 = s12[(s12['pvalue.adj.FDR'] < 0.05) & (s12['norm_foldChange'] > 0.5)]
s24 = s24[(s24['pvalue.adj.FDR'] < 0.05) & (s24['norm_foldChange'] > 0.5)]
s36 = s36[(s36['pvalue.adj.FDR'] < 0.05) & (s36['norm_foldChange'] > 0.5)]

d12 = d12.set_index('Unnamed: 0')
d24 = d24.set_index('Unnamed: 0')
d36 = d36.set_index('Unnamed: 0')

import numpy as np
int12 = set(s12.index) & set(d12.index)
s12int = s12[s12.index.isin(int12)].iloc[:, 10:]
s12int['l2fc'] = np.log2(s12int['norm_foldChange'])
s12int.sort_values(by='l2fc', key=abs, ascending = False).head(15)

import numpy as np

int24 = set(s24.index) & set(d24.index)

d24int = d24[d24.index.isin(int24)]

s24int = s24[s24.index.isin(int24)].iloc[:, 10:]
s24int['l2fc'] = np.log2(s24int['norm_foldChange'])
s24int.sort_values(by='l2fc', key=abs, ascending = False).head(15).

d24int = d24[d24.index.isin(int24)]
d24int

int36 = set(s36.index) & set(d36.index)

d36int = d36[d36.index.isin(int36)]

s36int = s36[s36.index.isin(int36)].iloc[:, 10:]
s36int['l2fc'] = np.log2(s36int['norm_foldChange'])
s36int.sort_values(by='l2fc', key=abs, ascending = False).head(15)

set(s12.index) & set(d12.index)

len(set(s24.index) & set(d24.index))

len(set(s36.index) & set(d36.index))

print(d12.index.shape[0])
print(d24.index.shape[0])
print(d36.index.shape[0])

print(s12.index.shape[0])
print(s24.index.shape[0])
print(s36.index.shape[0])

s8 = pd.read_csv('/Users/sofialuk/Desktop/thesis/Calu3/calu3_desingle/desingle_calu3_12h.csv')
s12 = pd.read_csv('/Users/sofialuk/Desktop/thesis/Calu3/calu3_desingle/desingle_calu3_8h.csv')

d8 = pd.read_csv('/Users/sofialuk/Desktop/thesis/Calu3/calu3_deseq/Calu3_DEGs_8h.csv')
d12 = pd.read_csv('/Users/sofialuk/Desktop/thesis/Calu3/calu3_deseq/Calu3_DEGs_12h.csv')


d12 = d12[(d12['padj'] < 0.05) & (d12['log2FoldChange'] > 0.5)]
d8 = d8[(d8['padj'] < 0.05) & (d8['log2FoldChange'] > 0.5)]

s12 = s12[(s12['pvalue.adj.FDR'] < 0.05) & (s12['norm_foldChange'] > 0.5)]
s8 = s8[(s8['pvalue.adj.FDR'] < 0.05) & (s8['norm_foldChange'] > 0.5)]

d12 = d12.set_index('Unnamed: 0')
d8 = d8.set_index('Unnamed: 0')

print('12:', len(set(s12.index) & set(d12.index)))
print('8:', len(set(s8.index) & set(d8.index)))

print(s12.index.shape[0])
print(s8.index.shape[0])

print(d12.index.shape[0])
print(d8.index.shape[0])

cell cycle regulators  (SPDYE8, ), genes involved in metabolism (PDPR2P, )
set(s12.Type.values)

d8[d8.index.isin(int8)].sort_values(by='log2FoldChange', key=abs, ascending=False).head(15)

s8int = s8[s8.index.isin(int8)].iloc[:, 10:]

import numpy as np
int12 = set(s12.index) & set(d12.index)
s12int = s12[s12.index.isin(int12)].iloc[:, 10:]
s8int['l2fc'] = np.log2(s8int['norm_foldChange'])
s12int['l2fc'] = np.log2(s12int['norm_foldChange'])
s12int.sort_values(by='l2fc', key=abs, ascending = False).head(15)

s8int.sort_values(by='chi2LR1', key=abs, ascending = False).head(15)
