==> 53_mirna_enrichr.ipynb <==
import os
import pandas as pd
import numpy as np
import mygene
import gseapy as gp
import gseapy.plot as gseaplot
import matplotlib.pyplot as plt

from gseapy import enrichment_map
import networkx as nx

# %matplotlib inline
# %config InlineBackend.figure_format='retina' # mac
%load_ext autoreload
%autoreload 2
import pandas as pd
import gseapy as gp
import matplotlib.pyplot as plt

from gseapy import Biomart
bm = Biomart()

from gseapy import gseaplot, heatmap

def enrich_genes(top_ranked_genes):

    enr_Reactome_up = gp.enrichr(gene_list=top_ranked_genes,
                                 gene_sets=['Reactome_2022'],
                                 organism='Human', 
                                 cutoff=0.5)
    
    enr_GOBP_up = gp.enrichr(gene_list=top_ranked_genes,
                             gene_sets=['GO_Biological_Process_2021'],
                             organism='Human', 
                             cutoff=0.5)
    
    enr_GOMF_up = gp.enrichr(gene_list=top_ranked_genes ,
                             gene_sets=['GO_Molecular_Function_2021'],
                             organism='Human', 
                             cutoff=0.5)
    
    
    # Write results to CSV files
    #file_name = os.path.basename(file_path).split('.')[0]
    #enr_Reactome_up.results.to_csv(f'{file_name}_enr_Reactome.csv')
    
    return enr_Reactome_up.results, enr_GOBP_up.results, enr_GOMF_up.results

targets4 = pd.read_csv('targets4.csv')
targets12 = pd.read_csv('targets12.csv')
targets24 = pd.read_csv('targets24.csv')

targets24

mirnas = set.union(set(targets24['mirna']), set(targets4['mirna']), set(targets24['mirna']))

targets24

mirnas = list(set(set(targets24['mirna'])))

mirnas[1]

targets24[targets24['mirna'].str.contains(mirnas[0])]

miRNA_targets = {}

# Iterate over each miRNA and find the corresponding targets
for mirna in range(len(mirnas)):
    # Filter the DataFrame based on the miRNA number in the target gene column
                   
    targets = targets24[targets24['mirna'].str.contains(mirnas[mirna])]
    miRNA_targets[f'miR-{mirnas[mirna]}'] = targets

# Create a Pandas Series from the miRNA_targets dictionary
miRNA_targets

mirnas = list(miRNA_targets.keys())

# Filter out the keys with empty Series values
miRNA_targets_filtered = {k: v for k, v in miRNA_targets.items() if not v.empty}

miRNA_targets_filtered

enr_Reactome_up, enr_GOBP_up, enr_GOMF_up = enrich_genes(miRNA_targets['miR-hsa-miR-374a-5p']['symbol'].dropna())

curr_mirna = 'miR-hsa-miR-374a-5p'
globals()[f"enr_Reactome_up_{curr_mirna.replace('-', '_')}"] = enr_Reactome_up

miRNA_targets_filtered.keys()

enr_Reactome_up, enr_GOBP_up, enr_GOMF_up = enrich_genes(miRNA_targets['miR-hsa-miR-29b-3p']['symbol'].dropna())
    
enr_Reactome_up_key = f"enr_Reactome_up_{curr_mirna.replace('-', '_')}"
globals()[enr_Reactome_up_key] = enr_Reactome_up
result_dict = {'miR-hsa-miR-29b-3p': globals()[enr_Reactome_up_key][:20]}

result_dict

enr_Reactome_up, enr_GOBP_up, enr_GOMF_up = enrich_genes(miRNA_targets['miR-hsa-let-7a-3p']['symbol'].dropna())
    
enr_Reactome_up_key = f"enr_Reactome_up_{curr_mirna.replace('-', '_')}"
globals()[enr_Reactome_up_key] = enr_Reactome_up
result_dict = {'miR-hsa-let-7a-3p': globals()[enr_Reactome_up_key][:20]}

result_dict

miRNA_targets_filtered.keys()

for curr_mirna in list(miRNA_targets_filtered.keys()):
    # Call the enrich_genes function and store the output in variables with dynamic names
    enr_Reactome_up, enr_GOBP_up, enr_GOMF_up = enrich_genes(miRNA_targets[curr_mirna]['symbol'].dropna())
    
    # Store the enriched gene lists in the globals() dictionary
    enr_Reactome_up_key = f"enr_Reactome_up_{curr_mirna.replace('-', '_')}"
    globals()[enr_Reactome_up_key] = enr_Reactome_up
    result_dict[curr_mirna] = globals()[enr_Reactome_up_key][:20]
    
print(result_dict)

import pickle

with open('dict_of_dfs.pickle', 'wb') as f:
    pickle.dump(result_dict, f)

result_dict

import pandas as pd

# Create an Excel writer object
writer = pd.ExcelWriter('enr_24hrs.xlsx')

# Iterate over the dictionary items
for key, value in result_dict.items():
    # Save each nested table as a separate sheet in the Excel file
    value.to_excel(writer, sheet_name=key, index=False)

# Save the Excel file
writer.save()
